buildable VSC

+++ examples/stm32/src/bsp.c
@@ -3,6 +3,8 @@
 #include <stddef.h>
 #include <stdio.h>
 #include <sys/time.h>
+#include <sys/stat.h>
+
 #include "stm32h5xx_ll_icache.h"
 #include "stm32h5xx_ll_pwr.h"
 #include "stm32h5xx_ll_crs.h"
@@ -324,7 +326,7 @@ void bsp_init_decoder(void) {
   NVIC_EnableIRQ(TIM15_IRQn);
 
   LL_TIM_InitTypeDef TIM_InitStruct = {0};
-  TIM_InitStruct.Prescaler = SystemCoreClock / 1000000;
+  TIM_InitStruct.Prescaler = (uint16_t)(SystemCoreClock / 1000000);
   TIM_InitStruct.CounterMode = LL_TIM_COUNTERMODE_UP;
   TIM_InitStruct.Autoreload = 65535;
   TIM_InitStruct.ClockDivision = LL_TIM_CLOCKDIVISION_DIV1;
@@ -367,7 +369,7 @@ void bsp_init_command_station(void) {
   NVIC_EnableIRQ(TIM15_IRQn);
 
   LL_TIM_InitTypeDef TIM_InitStruct = {0};
-  TIM_InitStruct.Prescaler = SystemCoreClock / 1000000;
+  TIM_InitStruct.Prescaler = (uint16_t)(SystemCoreClock / 1000000);
   TIM_InitStruct.CounterMode = LL_TIM_COUNTERMODE_UP;
   TIM_InitStruct.Autoreload = 0;
   TIM_InitStruct.ClockDivision = LL_TIM_CLOCKDIVISION_DIV1;
